backend/requirements.txt
flask
werkzeug
langchain
langchain-google-genai
langgraph
langchain-experimental
pandas
matplotlib
pytesseract
chromadb
pillow
streamlit
gunicorn
flask-sqlalchemy
pyjwt
